Add files via upload
Roman Numerals
This a program that prompts the user to enter a number within the range of 1 through 10. The program should display the Roman numeral version of that number. If the number is outside the range of 1 through 10, the program should display an error message. The following table shows the Roman numerals for the numbers 1 through 10:
          
          
          
          
            
              
              Number Roman Numeral1 I2 II3 III4 IV5 V6 VI7 VII8 VIII9 IX10 X