Average Rainfall
This program uses nested loops to collect data and calculate the average rainfall over a period of years. The program first asks for the number of years. The outer loop will iterate once for each year. The inner loop will iterate twelve times, once for each month. Each iteration of the inner loop will ask the user for the inches of rainfall for that month. After all iterations, the program will display the number of months, the total inches of rainfall, and the average rainfall per month for the entire period