Add files via upload
This is a program that keeps names and email addresses in a dictionary as key-value pairs. The program would display a menu that lets the user look up a person's email address, add a new name and email address, change an existing email address, and delete an existing name and email address.